+++ src/components/HeroesElements/index.tsx
@@ -32,6 +32,18 @@ export const HeroesElements: React.FC<Props> = (_props) => {
         role='One-time Sponsor'
         classes='p-10'
       />
+      <Card
+        imgUrl='https://avatars.githubusercontent.com/u/49100352?v=4'
+        link='https://github.com/AKP2401'
+        name='Aadharsh K Praveen'
+        role='One-time Sponsor'
+      />
+      <Card
+        imgUrl='https://avatars.githubusercontent.com/u/9942725?v=4'
+        link='https://github.com/krishhandro'
+        name='krishh'
+        role='One-time Sponsor'
+      />
 
       <Sponsors />