+++ src/components/Tooltip.tsx
@@ -18,10 +18,10 @@ const Tooltip: React.FC<Props> = (props) => {
       onMouseLeave={() => setShowTooltip(false)}
       onClick={() => setShowTooltip(!showTooltip)}>
       {showTooltip && (
-        <div className='absolute w-96 top-1/2 left-1/2  transform -translate-x-1/2 translate-y-4  shadow'>
+        <div className='absolute w-56 sm:w-96 top-1/2 left-1/2  transform -translate-x-1/2 translate-y-4  shadow'>
           <div className='flex max-w-xs flex-col items-center shadow-lg'>
             <div className='clip-bottom h-2 w-4 bg-black' />
-            <div className='rounded-xl bg-black p-2 text-[14px] sm:text-md text-white'>
+            <div className='rounded-xl bg-black p-2 text-[9px] sm:text-md text-white'>
               {props.tooltip}
             </div>
           </div>